Make greybus_messages.h public
- Will be needed for use in tests.
